feat: [#540] Remove unnecessary code
