--- utils.go
@@ -1,28 +1,0 @@
-package postgres
-
-import (
-	"strings"
-
-	"github.com/goravel/framework/errors"
-)
-
-func parseSchemaAndTable(reference, defaultSchema string) (string, string, error) {
-	if reference == "" {
-		return "", "", errors.SchemaEmptyReferenceString
-	}
-
-	parts := strings.Split(reference, ".")
-	if len(parts) > 2 {
-		return "", "", errors.SchemaErrorReferenceFormat
-	}
-
-	schema := defaultSchema
-	if len(parts) == 2 {
-		schema = parts[0]
-		parts = parts[1:]
-	}
-
-	table := parts[0]
-
-	return schema, table, nil
-}
