--- utils_test.go
@@ -1,30 +1,0 @@
-package postgres
-
-import (
-	"testing"
-
-	"github.com/goravel/framework/errors"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestParseSchemaAndTable(t *testing.T) {
-	tests := []struct {
-		reference      string
-		defaultSchema  string
-		expectedSchema string
-		expectedTable  string
-		expectedError  error
-	}{
-		{"public.users", "public", "public", "users", nil},
-		{"users", "goravel", "goravel", "users", nil},
-		{"", "", "", "", errors.SchemaEmptyReferenceString},
-		{"public.users.extra", "", "", "", errors.SchemaErrorReferenceFormat},
-	}
-
-	for _, test := range tests {
-		schema, table, err := parseSchemaAndTable(test.reference, test.defaultSchema)
-		assert.Equal(t, test.expectedSchema, schema)
-		assert.Equal(t, test.expectedTable, table)
-		assert.Equal(t, test.expectedError, err)
-	}
-}
